fix(financial): RefundService uses int arithmetic (ADR-018)
FASE D: RefundService + FinancialIntegrityTest con semántica int.

CAMBIOS:

1. RefundService::createReversalEntry (aritmética entera):
   - ANTES: $ratio = $refundAmount / $originalTotal (float)
   - DESPUÉS: $reversedDebit = (int) round($originalCredit * $refundAmount / $originalTotal)
   - Ajuste: $diff !== 0 en lugar de abs($diff) > 0.001
   - Invariant: SUM(debits) === SUM(credits) exacto

2. RefundService::createSimpleReversalEntry:
   - ANTES: 'credit' => (float) $refund->amount
   - DESPUÉS: 'credit' => (int) $refund->amount

3. FinancialIntegrityTest:
   - Eliminados 3 casteos (float) innecesarios
   - Balance check: abs($debits - $credits) < 0.02 → $debits === $credits

REGRESSION CHECK:
✅ FASE A (MoneyMigrationTest) 7/7
✅ FASE B (OrderTaxCalculationTest) 5/5
✅ FASE C (PaymentEntitiesTest) 11/11
✅ FASE C (BillIntegrityTest) 11/11
✅ FASE C (SplitBillTest) 4/4
✅ FASE D (FinancialIntegrityTest) 10/10

Siguiente: FASE E (CashSessionIntegrityTest)

+++ tests/Feature/BillIntegrityTest.php
@@ -468,8 +468,9 @@ test('Split por montos personalizados (splitByAmounts)', function () {
     $bills = $this->billingService->splitByAmounts($order, [6000, 4000]);
 
     expect($bills)->toHaveCount(2);
-    expect((float) $bills[0]->total)->toBe(6000);
-    expect((float) $bills[1]->total)->toBe(4000);
+    // ADR-018: total es integer, comparación directa
+    expect((int) $bills[0]->total)->toBe(6000);
+    expect((int) $bills[1]->total)->toBe(4000);
 });
 
 // ═══════════════════════════════════════════════════
@@ -555,8 +556,9 @@ test('Bill se puede reconstruir desde Order + Payments', function () {
 
     $reconstructed->registerPaymentAmount((int) $paidFromPayments + (int) $tipFromPayments); // ADR-018
 
-    expect((float) $reconstructed->paid_amount)->toBe(5000)
-        ->and((float) $reconstructed->remaining_amount)->toBe(5000)
+    // ADR-018: paid_amount y remaining_amount son integer
+    expect((int) $reconstructed->paid_amount)->toBe(5000)
+        ->and((int) $reconstructed->remaining_amount)->toBe(5000)
         ->and($reconstructed->status)->toBe(BillStatus::PARTIAL);
 });
